+++ HeroesCup.Web/Models/Missions/MissionIdeaViewModel.cs
