+++ HeroesCup.Web/Models/Missions/StoryViewModel.cs
